Disable interactions prevent execution (#9)
* Prevent execution when target is outdated

* Split bootstrap and loader

* Removed disable clicks when calling native api
