+++ ReactViewResources/Loader/Internal/Task.ts
@@ -7,11 +7,11 @@
         this.taskPromise = new Promise<ResultType>((resolve) => this.resolve = resolve);
     }
 
-    public setResult(result?: ResultType) {
+    public setResult(result?: ResultType): void {
         this.resolve(result as ResultType);
     }
 
-    public get promise() {
+    public get promise(): Promise<ResultType> {
         return this.taskPromise;
     }
 }
