+++ ReactViewResources/ReactViewResources.csproj
@@ -59,14 +59,15 @@
     <None Remove="AMDLoader\tsconfig.json" />
     <None Remove="index.html" />
     <None Remove="Loader.js" />
+    <None Remove="Loader\Bootstrap.ts" />
     <None Remove="Loader\Internal\ComponentsRenderCache.ts" />
     <None Remove="Loader\Internal\Environment.ts" />
     <None Remove="Loader\Internal\ErrorHandler.ts" />
     <None Remove="Loader\Internal\InputManager.ts" />
     <None Remove="Loader\Internal\MessagesProvider.ts" />
     <None Remove="Loader\Internal\NativeAPI.ts" />
-    <None Remove="Loader\Internal\NativeObjectProxy.ts" />
     <None Remove="Loader\Internal\ResourcesLoader.ts" />
+    <None Remove="Loader\Internal\ViewPropertiesProxy.ts" />
     <None Remove="Loader\Internal\ViewsCollection.ts" />
     <None Remove="Loader\tsconfig.json" />
     <None Remove="Loader\_Loader.js" />
@@ -105,6 +106,10 @@
 
   <ItemGroup>
     <EmbeddedResource Include="node_modules\react-dom\umd\react-dom.production.min.js" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="node_modules\scheduler\umd\scheduler.production.min.js" />
   </ItemGroup>
 
   <ItemGroup>
